backend/requirements.txt
uvicorn==0.34.2
fastapi==0.115.12
pydantic-settings==2.9.1
langchain[openai]==0.3.24
langsmith==0.3.33
python-dotenv==1.1.0
prometheus-fastapi-instrumentator==7.0.0
prometheus-client==0.21.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3